Updates to Fast Enough For You Solo
First - Emil, Thank You I've been a huge fan of emilstabs for going on decades now, and this tab is no exception.  Picked it up again after a decade and gave it a more thorough listen, and propose the following changes:

Change #1 @ 3:29 = the riff is subtle but has more notes than previously indicated
Change #2 @ 3:46 = the timbre of that bend sounds like a B string; it's softer/deeper/smoother than the equivalent on the E string, and the finger position makes more sense in the context of the preceding and following riffs
Change #3 @ 3:47 = sounds like a whole step to me and makes more sense given the positioning

Cheers.