Update Possum.txt
See https://youtu.be/zM6ENtw9dOQ 3:40 to 4:55, I corrected a mistake in my intro tab according to this lesson and careful rewatching of Trey's playing in several live videos. The original intro tab (I assume by Eric Berman) is not only difficult to play but is not how Trey plays it at all, therefore deleted.