Update Tela.txt
Corrected chord names in the intro & the non-instrumental verses. This makes it easier to tell what is happening in the jazz verses later on when the song changes key and then changes back.

Added jazz verse chord progressions. These were previously omitted, but are different from the non-instrumental verse progression. 

Cleaned up the formatting a bit. Just added titles before the verses to make it easier to read.